refactor: extract useConnectionGate (#3501 Phase 1t.3)
